+++ install_windows.ps1
@@ -24,39 +24,46 @@ Write-Output "smap will be installed to $installPath"
 $tempDir = Join-Path $env:TEMP "installNpcap" 
 New-Item -ItemType Directory -Path $tempDir
 
-# 1. Download Npcap and install
-$url_npcap = "https://npcap.com/dist/$npcap.exe"
-$outputPath = Join-Path $tempDir "$npcap.exe"
-Invoke-WebRequest -Uri $url_npcap -OutFile $outputPath 
-Start-Process -FilePath $outputPath -Wait
+$npcap_installed = Read-Host "Have Npcap been installed and SDK configured(y or ..)"
+if ($npcap_installed -ne "y") {
 
-# 2. Get system architecture 
-$x64 = [environment]::Is64BitOperatingSystem
-$ARM64 = [environment]::IsArm64OperatingSystem
+    # 1. Download Npcap and install
+    $url_npcap = "https://npcap.com/dist/$npcap.exe"
+    $outputPath = Join-Path $tempDir "$npcap.exe"
+    Invoke-WebRequest -Uri $url_npcap -OutFile $outputPath
+    Start-Process -FilePath $outputPath -Wait
 
-# 3. Npcap SDK
-$url_npcap_sdk = "https://npcap.com/dist/$npcap_sdk.zip"
-$sdkZip = Join-Path $tempDir "$npcap_sdk.zip"
-$sdkFile = Join-Path $tempDir $npcap_sdk
-Invoke-WebRequest -Uri $url_npcap_sdk -OutFile $sdkZip
-Expand-Archive -Path $sdkZip -DestinationPath $sdkFile
+    # 2. Get system architecture
+    $x64 = [environment]::Is64BitOperatingSystem
+    $ARM64 = [environment]::IsArm64OperatingSystem
 
-if ($x64) {
-    Copy-Item "$sdkFile\Lib\x64\*" $libPath
-} 
-elseif ($ARM64) {
-    Copy-Item "$sdkFile\Lib\ARM64\*" $libPath
-}
-else {
-    Copy-Item "$sdkFile\Lib\Packet.lib" $libPath
-    Copy-Item "$sdkFile\Lib\wpcap.lib" $libPath
+    # 3. Npcap SDK
+    $url_npcap_sdk = "https://npcap.com/dist/$npcap_sdk.zip"
+    $sdkZip = Join-Path $tempDir "$npcap_sdk.zip"
+    $sdkFile = Join-Path $tempDir $npcap_sdk
+    Invoke-WebRequest -Uri $url_npcap_sdk -OutFile $sdkZip
+    Expand-Archive -Path $sdkZip -DestinationPath $sdkFile
+
+    if ($x64) {
+        Copy-Item "$sdkFile\Lib\x64\*" $libPath
+    }
+    elseif ($ARM64) {
+        Copy-Item "$sdkFile\Lib\ARM64\*" $libPath
+    }
+    else {
+        Copy-Item "$sdkFile\Lib\Packet.lib" $libPath
+        Copy-Item "$sdkFile\Lib\wpcap.lib" $libPath
+    }
 }
 
 # 4. Install with cargo
 cargo install --path . --root $installPath
 
 # 5. copy resource folder
-Copy-Item -Path $resourceFolder -Destination $installPath -Recurse -Force
+$keep_res_files = Read-Host "Do you need to keep the resource files (please confirm that all resource files are working properly) (y or ..)"
+if ($keep_res_files -ne "y") {
+    Copy-Item -Path $resourceFolder -Destination $installPath -Recurse -Force
+}
 
 # 6. clear
 Remove-Item $tempDir -Recurse -Force
